fixed cart to work with database
